Lazily evaluate custom prediction message
