+++ engine/lib/src/kiln/util/CMakeLists.txt
@@ -1,5 +1,9 @@
 add_subdirectory(concepts)
 
+target_sources(${PROJECT_NAME} PRIVATE
+        contract_details.cpp
+)
+
 target_sources(${PROJECT_NAME} PUBLIC
         FILE_SET HEADERS
         BASE_DIRS ${PROJECT_SOURCE_DIR}/src
@@ -7,7 +11,7 @@ target_sources(${PROJECT_NAME} PUBLIC
         Any.hpp
         concepts.hpp
         contract_macros.hpp
-        contracts.hpp
+        contract_details.hpp
         Deallocator.hpp
         Dummy.hpp
         Function.hpp
