+++ engine/lib/src/kiln/util/contract_details.cpp
@@ -1,0 +1,29 @@
+#include "contract_details.hpp"
+
+#include <fmt/color.h>
+
+namespace kiln::util::internal {
+
+auto PreconditionViolation::print() const -> void
+{
+    fmt::println("{}", fmt::styled(headline(), fmt::fg(fmt::color::orange_red)));
+    fmt::println(
+        "    {} {} ({}:{})",
+        fmt::styled("source:", fmt::fg(fmt::color::spring_green)),
+        m_location.file_name(),
+        m_location.line(),
+        m_location.column()
+    );
+    fmt::println(
+        "    {} {}",
+        fmt::styled("function:", fmt::fg(fmt::color::yellow_green)),
+        m_location.function_name()
+    );
+    fmt::println(
+        "    {} {}",
+        fmt::styled("message:", fmt::fg(fmt::color::cornflower_blue)),
+        fmt::styled(m_message, fmt::emphasis::bold)
+    );
+}
+
+}   // namespace kiln::util::internal
