+++ engine/lib/src/kiln/util/contract_macros.hpp
@@ -2,9 +2,11 @@
 
 #include <source_location>
 
-#include "kiln/util/contracts.hpp"
+#include "kiln/util/contract_details.hpp"
 
-#define PRECOND(condition, ...)                                                           \
-    ::kiln::util::assert_precondition(                                                      \
-        condition, #condition, std::source_location::current() __VA_OPT__(, ) __VA_ARGS__ \
-    )
+#define PRECOND(condition, ...)                                                            \
+    ((void)((!!(condition))                                                                \
+            || (::kiln::util::internal::print_precondition_message_and_break(              \
+                    #condition, std::source_location::current() __VA_OPT__(, ) __VA_ARGS__ \
+                ),                                                                         \
+                false)))
